finanzas

+++ pages/finanzas/ingresos.vue
@@ -21,6 +21,11 @@
           S/.{{ item.monto }}
         </template>
         <template v-slot:[`item.acciones`]="{ item }">
+          <v-chip
+            v-if="item.categoria === 'Prestamo'"
+            :color="returnEstado(item)"
+            @click="changeType(item)"> {{ item.tipo === "Salida" ? 'Ingresar a ahorros' : 'Regresar a Salida' }}
+          </v-chip>
           <v-icon small color="blue" @click="editIngreso(item)">mdi-pencil</v-icon>
           <v-icon small color="red" @click="deleteIngreso(item.id)">mdi-delete</v-icon>
         </template>
@@ -36,7 +41,8 @@
   </v-card>
 </template>
 <script>
-import { getBoxs, deleteBox } from '../../services/boxServices';
+import Swal from "sweetalert2";
+import { getBoxs, deleteBox, updateBox } from '../../services/boxServices';
 import addIngreso from '../../components/finanzas/addIngreso.vue';
 export default {
   name: 'Ingresos',
@@ -62,6 +68,40 @@ export default {
     await this.fetchIngresos()
   },
   methods: {
+    async changeType(item) {
+      const confirm = await Swal.fire({
+        title: "¿Estás seguro que deseas devolver a Salida?",
+        // text: "Esta acción no se puede deshacer.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Sí, devolver",
+        cancelButtonText: "Cancelar"
+      });
+      if (confirm.isConfirmed) {
+        try {
+          await updateBox(item.id, {
+            tipo: 'Salida'
+          });
+          item.tipo = 'Salida'
+          await this.fetchIngresos()
+          Swal.fire("Movido", "El registro ha sido Movido.", "success");
+        } catch (error) {
+          console.error("Error al eliminar registro:", error);
+          Swal.fire("Error", "No se pudo Mover el registro.", "error");
+        }
+      }
+    },
+    returnEstado(item) {
+      let color = 'default' // Valor por defecto
+      if (item.tipo === "Salida") {
+        color = 'orange'
+      } else if (item.tipo === "Ingreso") {
+        color = 'purple'
+      }
+      return color
+    },
     async fetchIngresos(){
       try {
         const allBoxs = await getBoxs()
@@ -74,9 +114,28 @@ export default {
       this.selectedIngreso = {...ingreso}
       this.dialogComponent = true
     },
-    async deleteIngreso(id){
-      await deleteBox(id)
-      this.fetchIngresos()
+    async deleteIngreso(id) {
+      const confirm = await Swal.fire({
+        title: "¿Estás seguro que deseas eliminar?",
+        text: "Esta acción no se puede deshacer.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Sí, eliminar",
+        cancelButtonText: "Cancelar"
+      });
+
+      if (confirm.isConfirmed) {
+        try {
+          await deleteBox(id)
+          await this.fetchIngresos()
+          Swal.fire("Eliminado", "El registro ha sido eliminado.", "success");
+        } catch (error) {
+          console.error("Error al eliminar registro:", error);
+          Swal.fire("Error", "No se pudo eliminar el registro.", "error");
+        }
+      }
     },
     openDialog(){
       this.dialogComponent = true
